+++ src/apps/reader/Reader.entry.tsx
@@ -2,13 +2,10 @@ import React from "react";
 import { withConfig } from "../../core/utils/config";
 import { withUrls } from "../../core/utils/url";
 import { withText } from "../../core/utils/text";
-import Reader from "./Reader";
+import Reader, { ReaderType } from "./Reader";
 
-// interface ReaderEntryTextProps {}
-// interface ReaderEntryConfigProps {}
-
-const ReaderEntry: React.FC = () => {
-  return <Reader />;
+const ReaderEntry: React.FC<ReaderType> = ({ identifier, orderId }) => {
+  return <Reader identifier={identifier} orderId={orderId} />;
 };
 
 export default withConfig(withUrls(withText(ReaderEntry)));
