+++ src/apps/reader/Reader.stories.tsx
@@ -1,3 +1,4 @@
+import React from "react";
 import type { Meta, StoryObj } from "@storybook/react";
 import Reader from "./Reader.entry";
 
@@ -18,8 +19,11 @@ export default meta;
 
 type Story = StoryObj<typeof Reader>;
 
-export const Primary: Story = {
-  args: {
-    ...serviceUrlArgs
-  }
+export const WithIdentifier: Story = {
+  render: () => <Reader identifier="9788793681095" />
+};
+
+// Works only if the matrial is reserved
+export const WithOrderId: Story = {
+  render: () => <Reader orderId="97cb3a6f-e23b-41ad-97ca-e939541feba7" />
 };
